% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State code to be searched}

\item{year}{Year to be loaded}
}
\value{
A map with points of every accident for the state.num indicated for the year specified.
}
\description{
This function plots accidents in a specific year for a state in a map.
If the state code doesn't exists for the specified year, the function stops with an error.
It this functions from other packages:
		dplyr::filter
		maps::map
	graphics::points
}
\examples{
fars_map_state(37,2015)

}
