% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with the years to be processed.}
}
\value{
A data frame with the
}
\description{
This function load and summarize data from the file that corresponds to every year especified in the parameter.
It validates if every file exists and shows an error if it doesn't exists.
It uses mutate and select functions from dplyr package
}
\examples{
\dontrun{
fars_read_years(c(2013,2014,2015))
}

}
\seealso{
\code{\link{fars_read}}

\code{\link{make_filename}}
}
