% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with the years to be processed.}
}
\value{
A data frame with the summary
}
\description{
This function returns the quantity of incidets per month/year for every year especified in the parameter.
It uses bind_rows, group_by and summarize functions from dplyr package.
It uses spread function from tidyr package.
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
