% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year to be used as part of the name. It could be a string but it should be converted to a numeric.}
}
\value{
string containing the constructed filename
}
\description{
This create a the filename to be assigned to.
It doesn't validate if the parameter is a valid integer or even a valida number of year.
}
\examples{
fn <- make_filename(2015)

}
